+++ src/main/java/team/hackerping/nanuri/chat/application/command/ChatRoomCommand.java
@@ -1,4 +1,4 @@
-package team.hackerping.nanuri.chat.application.dto;
+package team.hackerping.nanuri.chat.application.command;
 
 public class ChatRoomCommand {
     public record Create(
